Fix: Bug in OTA PALs for Microchip & ESP
git issue #90

aws_ota_pal.c for Microchip and ESP did not check if valid
code verification certificate handle was found, and proceeds to
malloc a 0-length buffer for a non-existent certificate.  This
caused OTA code verification to fail for OTAs relyng on certificate
located in aws_ota_codesigner_certificate.h.

This fix makes sure the prvGetCertificate checks if the certificate
handle is valid before mallocing memory for it, and implements a
check for a non-null certificate value in prvPAL_ReadAndAssumeCertificate.

Small updates to factory_image_generator.py script to make it agnostic of
Python version 2 or 3.