Disable CONFIG_PTHREAD_MUTEX_TIMEDLOCK in ESP-IDF
Why:
This is to prepare the merge of CMake for ESP32. This macro needs to be
disabled since it brings dependency on our POSIX implementation, which
has to be disabled in ESP32 for now because we haven't come up with a
solution how vendor partners should support the POSIX interface.