fix: Use accessor macro to access list member
FreeRTOS_TCP_WIN.c accessed the container member of the list item
directly instead of using the provided accessor macro for the same. The
container variable name was changed from pvContainer to pxContainer
which resulted in build failure. To ensure backward compatibility,
pxContainer was replaced with pvContainer using preprocessor if
configENABLE_BACKWARD_COMPATIBILITY was set to 1.

This commit removes the direct access and replaces it with the provided
accessor macro. As a result, it is no longer needed to set
configENABLE_BACKWARD_COMPATIBILITY to 1.

Fix issue: https://github.com/aws/amazon-freertos/issues/366

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>