Merge/ble beta master3 (#387)
* style: Uncrustify all files

Many files in our repository were not uncrustified (atleast according to
the uncrustify config we provide) which resulted in a failed pre-commit
check if any change was made to any such file. This commit uncrustifies
all the files. This commit also disables Hungarian notation check.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* Convert errors in TLS wrapper to negative error codes (#358)

* Convert positive error codes in TLS wrapper to negative error codes

TLS library states that it returns negative error codes, but PKCS #11
library returns positive error codes.  Some positive PKCS #11 error codes
were propagated out of TLS, breaking API contract and occasionally causing
error code to be interpreted as number of bytes sent or received.

This commit also fixes a threading bug in the Windows PKCS #11 PAL layer
when reading files, and cleans up formatting on an error print statement in
the sockets tests.

* Fix some issues in quick start script under Windows

- We cannot delete read-only file under Windows, need to change
  permission first.
- cert_id_file is open without being closed.

* Add the 'off_t' POSIX type and ensure that 'sys/types.h' is included in 'unistd.h' so that the 'useconds_t' type is known to be known. (#350)

* Add the 'off_t' POSIX type and ensure that 'sys/types.h' is included in 'unistd.h' so that the 'useconds_t' type is known to be known.

* Prefix the 'sys/types.h' include with 'FreeRTOS_POSIX' to ensure that the correct 'sys/types.h' file is included.

* CMake support for ESP-IDF (#359)

* ESP-IDF : CMake build system files and scripts updated

* Disable CONFIG_PTHREAD_MUTEX_TIMEDLOCK in ESP-IDF

Why:
This is to prepare the merge of CMake for ESP32. This macro needs to be
disabled since it brings dependency on our POSIX implementation, which
has to be disabled in ESP32 for now because we haven't come up with a
solution how vendor partners should support the POSIX interface.

* Fix ESP32 build problem with IDF update

* Why: TLS_Send does not handle the error condition when space is not avaiable on disk.

Related Issues: https://github.com/aws/amazon-freertos/issues/275

* esp32: support for power management framework

* fix: Use accessor macro to access list member

FreeRTOS_TCP_WIN.c accessed the container member of the list item
directly instead of using the provided accessor macro for the same. The
container variable name was changed from pvContainer to pxContainer
which resulted in build failure. To ensure backward compatibility,
pxContainer was replaced with pvContainer using preprocessor if
configENABLE_BACKWARD_COMPATIBILITY was set to 1.

This commit removes the direct access and replaces it with the provided
accessor macro. As a result, it is no longer needed to set
configENABLE_BACKWARD_COMPATIBILITY to 1.

Fix issue: https://github.com/aws/amazon-freertos/issues/366

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* fix: WinPCAP network interface selection (#373)

* feat: Update FreeRTOS kernel to version 10.2.0

This commit updates the FreeRTOS kernel to version 10.2.0 to bring it in
sync with the SVN repo. The major changes in this kernel version are
support for RISC-V and ARMv8M.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* fix: Fix warning portHAS_STACK_OVERFLOW_CHECKING not defined

portHAS_STACK_OVERFLOW_CHECKING was getting defined too late before
being used in portable.h for the platforms that do not have stack
overflow checking registers. This commit ensures that it is defined
before it is used.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* bugfix to "ARG FV Finding: FreeRTOS_setsockopt() returns wrong error codes".
TT E039512548

* fix: Fix duplicate defined FreeRTOS hooks in Microchip.

* Took rtos_hooks.c out of the compilation to match current test project.

* SAVE

* Removing deprecated OTA file