Feature/ble beta merge (#412)
* style: Uncrustify all files

Many files in our repository were not uncrustified (atleast according to
the uncrustify config we provide) which resulted in a failed pre-commit
check if any change was made to any such file. This commit uncrustifies
all the files. This commit also disables Hungarian notation check.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* Convert errors in TLS wrapper to negative error codes (#358)

* Convert positive error codes in TLS wrapper to negative error codes

TLS library states that it returns negative error codes, but PKCS #11
library returns positive error codes.  Some positive PKCS #11 error codes
were propagated out of TLS, breaking API contract and occasionally causing
error code to be interpreted as number of bytes sent or received.

This commit also fixes a threading bug in the Windows PKCS #11 PAL layer
when reading files, and cleans up formatting on an error print statement in
the sockets tests.

* Fix some issues in quick start script under Windows

- We cannot delete read-only file under Windows, need to change
  permission first.
- cert_id_file is open without being closed.

* Add the 'off_t' POSIX type and ensure that 'sys/types.h' is included in 'unistd.h' so that the 'useconds_t' type is known to be known. (#350)

* Add the 'off_t' POSIX type and ensure that 'sys/types.h' is included in 'unistd.h' so that the 'useconds_t' type is known to be known.

* Prefix the 'sys/types.h' include with 'FreeRTOS_POSIX' to ensure that the correct 'sys/types.h' file is included.

* CMake support for ESP-IDF (#359)

* ESP-IDF : CMake build system files and scripts updated

* Disable CONFIG_PTHREAD_MUTEX_TIMEDLOCK in ESP-IDF

Why:
This is to prepare the merge of CMake for ESP32. This macro needs to be
disabled since it brings dependency on our POSIX implementation, which
has to be disabled in ESP32 for now because we haven't come up with a
solution how vendor partners should support the POSIX interface.

* Fix ESP32 build problem with IDF update

* Why: TLS_Send does not handle the error condition when space is not avaiable on disk.

Related Issues: https://github.com/aws/amazon-freertos/issues/275

* esp32: support for power management framework

* fix: Use accessor macro to access list member

FreeRTOS_TCP_WIN.c accessed the container member of the list item
directly instead of using the provided accessor macro for the same. The
container variable name was changed from pvContainer to pxContainer
which resulted in build failure. To ensure backward compatibility,
pxContainer was replaced with pvContainer using preprocessor if
configENABLE_BACKWARD_COMPATIBILITY was set to 1.

This commit removes the direct access and replaces it with the provided
accessor macro. As a result, it is no longer needed to set
configENABLE_BACKWARD_COMPATIBILITY to 1.

Fix issue: https://github.com/aws/amazon-freertos/issues/366

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* fix: WinPCAP network interface selection (#373)

* feat: Update FreeRTOS kernel to version 10.2.0

This commit updates the FreeRTOS kernel to version 10.2.0 to bring it in
sync with the SVN repo. The major changes in this kernel version are
support for RISC-V and ARMv8M.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* fix: Fix warning portHAS_STACK_OVERFLOW_CHECKING not defined

portHAS_STACK_OVERFLOW_CHECKING was getting defined too late before
being used in portable.h for the platforms that do not have stack
overflow checking registers. This commit ensures that it is defined
before it is used.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* bugfix to "ARG FV Finding: FreeRTOS_setsockopt() returns wrong error codes".
TT E039512548

* fix: Fix duplicate defined FreeRTOS hooks in Microchip.

* Took rtos_hooks.c out of the compilation to match current test project.

* feat: Add checksums file containing checksums of our files

This file is used by IDT to check the integrity of the code.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* fix: Free the memory allocated by pcap_compile().

* pcap_freecode() is used to free up allocated memory pointed to by a bpf_program struct generated by pcap_compile(3PCAP) when that BPF program is no longer needed, for example after it has been made the filter program for a pcap structure by a call to pcap_setfilter(3PCAP).

* fix: Delete deprecated in-house developd files from third_party/win_pcap

* These arch.c and netif.h in lib/third_party/win_pcap are not compiled or used
  in the pc/windows project. These files are left over from an older implementation
  of FreeRTOS+TCP.

* fix: Remove the warning of unused variable xTaskStatus

There was a warning generated on System Workbench of -Wunused-variable.
The reason being it falls under the preprocessor block of INCLUDE_uxTaskGetStackHighWaterMark.

* fix: Make vLoggingPrint() for WinSim scheduler safe.

* printf and fflush are not scheduler safe operations.
* vLoggingPrint provides a variation of the WinSim console logger that omits timestamps.

* docs: Fix spelling of priority in comments

This was pointed out in the PR:
https://github.com/aws/amazon-freertos/pull/320

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* fix: pcap_freecode() was in the wrong spot.
* pcap_freecode() needs to be called anytime pcap_compile succeeds.

* feat: Use SHA256 for file checksums instead of SHA1

This file contains checksums for all AFR files and is used by IoT Device
Tester for checking the integrity of the code.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* esp-idf: fix regression in interrupt allocation routines for critical section API usage

* docs: Fix README.md style for Windows Simulator.

The Windows Simulator sections was inconsistently styled from the other platform's sections.

* Mahavirj bugfix/esp multiple fixes (#404)

* lib/wifi: espressif: fix issue with WiFi configuration for non-null strings

(cherry picked from commit 7ac95f20522f66b95e612c113e54b3a506635b66)

* esp32: lib/FreeRTOS-Plus-TCP: do not send eNetworkDownEvent to stack if interface is already down

(cherry picked from commit 7c3ad5a678b31f86c5b929c34331fcafb6eb8be5)

* mbedtls: configurable options for controlling dynamic memory allocations

Allows to reduce memory per TLS connection, default values set are 8K for
incoming fragment and 4K for outcoming fragment.

* lib/wifif: espressif: fix scan failure under certain disconnect scenarios

* lib/third_party: update ESP-IDF to latest v3.1.3 release

Detailed release notes at https://github.com/espressif/esp-idf/releases/tag/v3.1.3

* NetworkInterface: Check interface state before sending packet to WiFi driver

(Open question here is: even though state of network interface is conveyed to TCP stack
it keeps on sending requests for TX to WiFi driver, should this be handled in TCP stack?)

* Saving config

* fix: Fix link failure with latest XC32 compiler on Microchip

The latest XC32 compiler (version 2.15) does not allow multiple
definitions while the previous versions allowed it. As a result, using
the latest compiler results in linker errors. We should remove
multiple definitions but allowing explicitly for now to fix the
build failure with the latest compiler.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>

* Revert aws_wifi.c to master branch version.

* Fixing WIFI

* Fixing demo runner