Add aws_mbedtls_pkcs11-level tracking of label-handle lookup
In order to not change the PKCS #11 PAL layer, a mapping of
handle to label must be maintained in the PKCS code at the
aws_mbedtls_pkcs11.c level.  This commit adds a an array of
handle-label pairs to the PKCS #11 module structure, as well as
functions to add, search, and destroy these entries.