Add configPRECONDITION to FreeRTOS.h
I introduce `configPRECONDITION` which is resolved to `configASSERT` in the default case,
but allows finer control for dealing with an assertion in CBMC proofs.
There are a couple of implicit invariants, I will introduce as `configPRECONDTION` along with test harnesses to the code base.
Other invariants are explicit and keeps as `configASSERT`.
The idea is to resolve `configPRECONDITION` to `__CPROVER_assume` for proofs and `configASSERT` to `__CPROVER_assert` for proofs.