Fix out-of-bound read in the MQTT library
A malformed MQTT Publish packet received from the broker could cause
the parsing code to read out-of-bound - MQTT strings are prefixed with
a two byte length field which gives the number of bytes in the string
itself. The incorrect length could cause out-of-bound read. This change
ensures that such a malformed packet does not cause out-of-bound read.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>