Make the driver for Inventek WiFi module interrupt driven
The driver for the Inventek WiFi module provided by ST had support for
interrupts but it was written for CMSIS-RTOS. So far we were using the
driver in the polling mode thereby wasting unnecessary CPU cycles while
the module was busy working.

This change uses the CMSIS wrapper for FreeRTOS provided by ST to enable
the interrupt mode of the driver. As a result, we no longer will waste
any unnecessary CPU cycles while the module is busy working.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>