Make disconnect delay configurable.
Quoting from PR feedback --
The other consideration in original implementation is to try to wait for server response before disconnect. (In a sense that, if user does care about /rejected or /accepted response, the user needs to figure out a reasonable AWS_IOT_DEFENDER_WAIT_SERVER_MAX_SECONDS. Depending on which region the device is trying to connect to from where, this value could be different.)

Since we do have a plan for updating the defender report scheduling, I'm going to revert the change to use macro here. But the macro value will be updated to zero. The follow-up items would be:

- update scheduling.
- update demo, to actually show how callback works.