Fix WiFi test failures/hangs for Mediatek
Mediatek DHCP code creates a DCHP task in dhcpd_start function which
creates a LWIP socket and listens on it. dhcpd_stop function deletes
the DHCP task and then closes the socket. When the socket is closed,
LWIP makes sure that if any task is waiting for the data to be received
on the socket, it gets unblocked. The following piece of code from LWIP
does that:

while( xTask != NULL )
{
    xTaskAbortDelay( xTask );
    taskYIELD();
    xTask = pxMailBox->xTask;
}

dhcpd_stop function first deletes the DHCP task and then closes the
LWIP socket. This creates a race condition –

 - DHCP task calls recv_from on the socket and gets blocked waiting for
   the data on the socket.
 - User task (test runner in our case) calls dhcpd_stop which deletes
   the DHCP task and then attempts to close the above socket. Since the
   DHCP task is still waiting for the data, the above mentioned while
   loop tries to unblock it - A deleted task cannot be unblocked/run
   and therefore the above while loop becomes an infinite loop.

This change fixes the dhcpd_stop function to first close the socket and
then delete the DHCP task.

WIFI_Disconnect implementation incorrectly assumes that "link up" means
that the WiFi module is connected to an AP. This causes problem because
AFQP_WIFI_ConnectAP_InvalidSecurityTypes test provides incorrect
security type WEP instead of WPA2 which results in auth error and leaves
the module in "link up" but disconnected state. The call to
WIFI_Disconnect in the setup function then tries to disconnect the already
disconnected module and enters an infinite wait for a callback which
never gets invoked. This change fixes the WIFI_Disconnect implementation
to check the connected state variable (which is set in the connection
successful callback) to determine whether or not the module is already
connected.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>