Remove vTaskDelete calls from demos
All the demos run in the task named runDemoTask which is created using
Iot_CreateDetachedThread. The use of this API requires that the task
must not call vTaskDelete which will result in a memory leak. Also, the
demo status string and the demo end marker string which are printed in
the runDemoTask function will not get printed if the runDemoTask is
deleted pre-maturely by calling vTaskDelete. This change removes the
vTaskDelete calls from greengrass and defender demos to ensure that the
runDemoTask is not deleted pre-maturely.

Also, the defender demo was not propagating the errors from APIs which
could result in a false success code from the defender demo. This change
fixes the demo to propagate the errors from the defender APIs.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>