Fix integer overflow in CBMC proofs caused by long buffer lengths.
In CBMC, a pointer is an object id followed by an offset into the
object.  The number of bits of the pointer used for the object id
limits the number of bits remaining for the offset, and hence
limits the size of the object.  If the top 7 bits are used for the
object id, then offests into the object must fit into the bottom
32-7 bits.