Simplify disabling public modules for a hardware platform
AFR public modules are the modules which are selectable by the user in
the FreeRTOS console (OCW). We need to disable some public modules for
some hardware platforms - for example, the Nordic board cannot support
HTTPS library because it does not have WiFi/Ethernet. We added a
function, namely afr_disable_public_modules, to disable public modules
for a hardware platform. Later we realized that this function is not
needed and a module (regardless of whether it is public or not) can be
disabled by just calling set(AFR_MODULE_<module> 0 CACHE INTERNAL "") in
the board's CMakeLists.txt.

This change removes the function afr_disable_public_modules and instead
uses set(AFR_MODULE_https 0 CACHE INTERNAL "") to disable HTTPS module
for the Nordic board.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>