Add delay before starting a demo
The test framework flashes the binary to a board and then opens the
serial  port for reading output. The code on the MCU starts running
after it is flashed and as a result it is possible that the test
framework may not see the "DEMO STARTING" tag. To address this, this
change adds delay before starting the demo and this delay is only added
if INSERT_DELAY_BEFORE_DEMO is defined - This ensures that no one else
sees this delay.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>