make_common_makefile now returns list
Previously, make_common_makefile.py would always write a Makefile to a
file. This commit makes the script return a list of lines, writing that
list to a file if the script is directly invoked. This is to prepare for
make_common_makefiles being called from another script, which will
append further lines to the makefile before writing it out to a file.