make_proof_makefiles now returns list
Previously, make_proof_makefiles.py would dump a Makefile in each proof
directory. This commit modifies it so that the function that does all
the work returns a list of makefile lines rather than writing the lines
to a file. The main() method then writes the lines to a file.

This is in preparation for an external script that will call into
make_proof_makefiles and get a list of makefile lines, and add to them
before writing out a final makefile