Dump complete Makefiles in each proof directory
prepare.py now walks the entire directory tree and dumps a
self-contained Makefile in each proof directory.

Previously, prepare.py dumped common recipes to the file
`proofs/Makefile.common`, and then dumped proof-specific defines in a
proof-specific Makefile that included the common Makefile. This commit
concatenates the proof-specific defines to the common recipes and dumps
that into each proof directory.

This is in preparation for a commit that will also customize the recipes
that get dumped into each proof directory.