Remove redundant loops from the same-socket/different-tasks test. (#1227)
The same-socket/different-tasks TCP group test executes the same four variations in a loop four times consecutively. The architecture of the test is such that the sending and receiving threads are synchronized at the end of each loop. On some embedded boards with Wi-Fi connectivity, a single iteration can take several minutes to complete. This change runs each variation once.