fix: iot_battery_temp() needs to return signed value (#1351)
iot_battery_temp() API needs to return signed, not unsigned value

Why:  Most of this is documentation changes, but there is one API
change.  While reviewing the API's, I found that iot_battery_temp()
was originally returning an unsigned value.  This needs to be
signed since temperatures can be below 0.