style: Documentation improvements and spelling fix in rtc APIs (#1391)
Documentation improvements as well as a spelling fix in the rtc APIs

Why:  While making documentation improvements in the rtc APIs, we
found that there was a spelling error in one of the enums.  That
has been fixed, along with clarify what return values are expected
for all of the API's and what datatypes are expected on the ioctl
calls.