Delete CSDK maintained docs.
This is done because the documentation for CSDK libraries has been repeated,
but not actively maintained. These docs were here simply for developer
convenience in doxygen links, despite having some outdated information. The
README.md and mainpage are updated to note where the docs for the CSDK shared
libraries reside.

Other updates:
* Update doc/README.md for correct CSDK api ref link.
* Update doc README.md to specify where CSDK doxygen docs are maintained.
* Synchronize the doxygen content of the platform layer in AFR to CSDK.
  The IoT platform is kept because Amazon FreeRTOS has ported it.
* Update main layout and _LIBRARY_LOG_X macros for sync with CSDK.