test: Fix bug in testing of iot_uart APIs (#1432)
Fixing bug in ioctl testing of iot_uart APIs

Why:  Existing test was trying to test for config values changing
when calling ioctl xUartSetConfig without first setting the value
to a known state.  This test would falsely fail if the value being
changed to happen to match with the previous value.  The fix is
to set the config to a set of known values first, and then one by
one change each parameter, and confirm.