Don't use bogus loop for CBMC unwindset
The --unwindset flag to CBMC requires an argument, but many proofs don't
need an unwindset. This commit ensures that --unwindset is not used at
all if a proof's unwindset is empty, instead of using a bogus loop ID as
an argument. This makes it easier to see if a real loop ID was
misspelled if/when CBMC adds such a feature.