Allow proofs to use --nondet-static{,-except}
Proof-writers may now set either the `cbmc_nondet_static` variable to ON
(or OFF, which is the default), or add variables to the
`cbmc_nondet_static_exclude` list. This will insert a goto-instrument
pass that makes all static variables in the program start with an
unconstrained value, with the exception of any variable in the
`cbmc_nondet_static_exclude` list.

This has the effect that CBMC will not assume that static variables have
been initialized (e.g. to zero), but will check the program for all
possible initial values of static variables.