Do not replace / with \ for cbmc-viewer on Windows
cbmc-viewer expects pathnames passed to the --srcexclude flag to be
separated by forward-slashes, no matter what the platform is. The
make_common_makefile.py script was nevertheless replacing the
forward-slash in Makefile.template with a backslash when run on Windows,
meaning that cbmc-viewer took much longer than needed (because it was
scanning excluded directories).

This commit provides a mechanism for having forward-slashes in
Makefile.template that do not get replaced with backslash on Windows.