Fix Infinite loop in BLE disconnect
Shadow demo and MQTT demo on Nordic were not completing on Nordic.
The root cause was that the demo was completing but it was ending up in
an infinite loop in BLE disconnect logic and thus blocking the lower
priority logging task . This prevented logs of demo completion to be not
printed.
Solution is done by exiting the infinite loop in a condition when there
is error.