Fix a dependency parsing bug in CMake from #1161
In previous PR when we add the feature to disable building demos and
test, we add a check to not search dependencies of aws_demos and
aws_tests target when demos or tests are disabled. But we forgot to also
move adding the dependencies to demos_base and tests_base inside the
check. As a result, the "dependencies" from previous for loop was added
to demos_base and tests_base.