Bringing changes for AFR Release 202002.00 (#1723)
* Fix a dependency parsing bug in CMake from #1161 (#1655)

In previous PR when we add a feature to disable building demos and
test, we add a check to not search dependencies of aws_demos and
aws_tests target when demos or tests are disabled. But we forgot to also
move adding the dependencies to demos_base and tests_base inside the
check. As a result, the "dependencies" from previous for loop was added
to demos_base and tests_base.

* Add board dir for cypress in CMake (#1684)

* Merge pull request #1700 from tgsong/ota_e2e_s3_cleanup

Only delete files uploaded in OTA E2E

* OTA e2e scripts: Replace dots with underscores (#1717)

* Update FreeRTOS-Kernel submodule branch to release-candidate

Update FreeRTOS-Kernel submodule branch from V10.2.1-convergence-FreeRTOS-Source
to release-candidate_10.3.0_rc3
commit id: d7550ef1a0560ceddac7c0b8bc0148272ac641a6

Co-authored-by: Tiangang Song <ts.whu@outlook.com>
Co-authored-by: Pavan Madehalli Ranganath <rangpava@amazon.com>