Explicit casting unsigned long to TimerHandle_t
The timer_t type when expanded results in an unsigned long in some Microchip compilers. The compiler does an implicit typecast (which is also bad) which we cannot get around, thus to suppress the warnings, we explicitly typecast the unsigned long to TimerHandle_t. TimerHandle_t is a struct pointer.