Change Default Winsim Test Port to 0
Currently the winsim test project is configured to network adapter 3. If this adapter is incorrect for your windows PC, the project will seemingly
get stuck after successfully opening the network port. This issue is very hard to debug, as the project appears to just be frozen. If we configure this
network port to 0, the test project will be unable to open the network adapter, and print a helpful error message that helps the user configure it to the right
network adapter.