Fix defeneder test build failure on multiple platforms (#1936)
Cypress - A missing file in tinybor resulted in linker error.
ESP32 -  Defender test files were not getting built in the make files.
Marvell: Marvell passes -ffreestanding flag to the compiler which
results in __STDC_HOSTED__ being defined to 0. In this case the
function used for printing a cbor value cbor_value_to_pretty is not
available and this results in a linker error. 

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>