Fix build failure for TI tests for cmake (#1950)
The Ninja generator combines linker and post build commands using '&&'
which results in a long string. Windows has a limit of 8191 characters
on the command line string length. We were hitting that limit for TI and
were getting "command string too long" error.

This commit updates the cmake files to force Ninja to use response files
which is a feature implemented for exactly this purpose and results in a
much smaller command string.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>