Make each config of a CBMC proof have unique name
Some CBMC proofs are built in several different 'configurations'. Prior
to this commit, the name of each of these configurations was the same as
the name of the original proof. This meant that CI would sometimes write
the proof artifacts of each of these configurations on top of each
other, since the artifacts are saved under a folder whose name is based
on the name of the proof.

This commit fixes this so that each configuration of each proof has a
unique name, formed by joining the name of the configuration to the name
of the proof with a double underscore---for example,
TaskIncrementTick__config_default.