Fix for IDT failing for internal FLASH
problem is deeper than this solution.
boot_set_pending() and boot_set_confirmed()
call boot_write_trailer() which checks alignment
values and we don't pass the test.

Ignoring the return values has been the way we
have done this since Q4 2019 when we first created
the AFR OTA solution.

We will address the underlying issue when we create
the middleware layer of FLASH+OTA utilities