AFRSDK-92
Move boot_set_pending() from "FileClose" to "Activate".

By looking at the new IDT 3.1 test failures, I reasoned that the
Activate call from the AFR OTA Agent was happening even though
the device PUBLISH to the MQTT server fails. I created a GitHub issue
for this with Amazon: https://github.com/aws/amazon-freertos/issues/2361

The basic issue is that the device does not reconnect with the MQTT
broker, even though there multiple failures PUBLISHing to the MQTT
broker from the OTA Agent. PING failures do cause the reconnect
to the MQTT broker. To help this:

Reduce the block size so that the amount of time it takes
to download the OTA Image is longer.
Reduce the KEEP_ALIVE time so that a PINGREQ happens before
the OTA Image completes.

This change is expected to fix these tests:
MQTT & HTPP OTATestDisconnectAndResume
MQTT & HTTP OTATestDisconnectCancelUpdate