Add script for running all proofs in parallel
This commit adds the run-all-proofs.sh script, which builds all proofs
in parallel using Litani. It also changes Makefile.template to ensure
that all goto-binaries are built out of source, to ensure that it is
safe to build proofs in parallel without the object files that they
depend on being overwritten.