Add public key support in the PKCS11 PAL for ST
This is needed for PKCS11 EC tests to pass when TLS is done on the host
itself as opposed to offloading to the WiFi module. In addition to
these changes, the following config changes are also needed to run EC
tests:
- Set pkcs11testEC_KEY_SUPPORT to 1 in iot_test_pkcs11_config.h.
- Set pkcs11testGENERATE_KEYPAIR_SUPPORT to 1 in iot_test_pkcs11_config.h.
- Set pkcs11configSUPPRESS_ECDSA_MECHANISM to 0 in iot_pkcs11_config.h.

This commit also fixes a bug in the iot_pkcs11_mbedtls.c where incorrect
label (public key/private key) was used to find the corresponding
object.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>