afr.cmake: Add BUILD_CLONE_SUBMODULES option
This option is similar to the option under the same name already used
in some submodules of the repository. If set to OFF, it skips automatic
submodule update during building. This is useful for focused builds
of particular targets, which use only a subset of submodules (which
then need to be updated externally). In particular in continuous
integration contexts, this can save build time/bandwidth.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>