Fix build combinations issues for defender demo (#2858)
* Fix missing freertos_plus_tcp_utils

This change brings the freertos_plus_tcp_utils module in when the
device_defender_demo_dependencies module is selected for the FreeRTOS
console.

* Dedupe defender demo CMakeLists

Moves the duplicated parts of the defender demo cmake configuration out
of the metrics collector conditional blocks.

* Remove explicit lwip dependency for metrics_collector

ESP32 doesn't use 3rdparty::lwip, and the explicit dependency was not
nessesary, so this change removes it from the configuration for lwip
metrics collection.

* Split freertos_plus_tcp_utils into seperate file

Build combinations were failing when freertos_plus_tcp_utils was in the
same cmake file as freertos_plus_tcp. This moves it to its own file with
the file name same as the module in order to fix the build combinations
issue. This also fixes the build failures on Marvell.

* Run IDE generator