Add a configurable delay in test runner before starting tests (#2950)
There is a race condition between the test runner executing on the device and the AWS IoT Device Tester gaining access to the device through serial connection on the host machine. The Device Tester is unable to detect beginning of tests on device if the device takes time to be available on the host machine OS.
This PR tackles the race condition problem by adding a configurable delay period before the start of the tests in the test runner so that the tests do not start executing before the device is available on the serial port of the host machine.