Add CLI input mode to test runner (#2955)
Add another solution ( separate from the configurable delay config added in PR #2950 ) for the problem of race condition between AWS IoT Device Tester (that runs on host machine) getting serial console access to device and the test runner executing tests on the device.

This change adds a CLI console feature to the test runner of waiting for a "start" command from the console (based on FreeRTOS+CLI library) to being execution of tests. This feature can be enabled with the AWS_TEST_RUNNER_ENABLE_CLI_INPUT config in the aws_test_runner_config.h file.