Fix hangups caused by blocking forever in network receive task when calling shutdown (#2986)
In some TCP/IP stacks like lwIP, shutting down a socket causes a pending recv to block forever. Therefore, IotNetworkAfr_Close is modified to wait for the network receive task to exit so that the socket can be shutdown safely without causing the socket to block forever if there are pending reads or writes being done from other tasks.