Retry subscribing to topic filter when SUBACK not received (#2983)
For Shadow, Defender, and Jobs demos, the function SubscribeToTopic would have returned successfully even though the SUBACK wasn't received. This is due to the fact that MQTT_ProcessLoop returns success even when there is no data to receive and does not wait to receive any packet type. This fixes it so as to add the logic from the MQTT Mutual Auth demo to retry the subscription whenever the SUBACK is not received within the timeout and also returns failure once all attempts are exhausted.