Move code signing cert. to demo/test config files
The OTA PAL uses a macro to define a code signing certificate. This
change moves the macro to a location common to the demo and test files.
The macro is now being defined in the ota_config.h file via either a
testing config file or a demo config file that is being included.

This change included the following sub-changes:
* Update the OTA library CMake target to depend on the OTA MCU port.
* Update the macro name of the code signing certificate.
* Update the ESP32 OTA MCU port dependencies.
* Add the ota_demo_config.h include to the ota_config.h for both esp32
and winsim demos.
* Update the esp32 and winsim ota_config.h files that are part of the
test projects to include the test config file.
* Update the ota_pal.c ports for esp32 and winsim to include
ota_config.h.
* Clean up artifacts of a previous git merge.