Update the application version to only be used in OTA demos (#3158)
Remove demo app version related code
* Delete the aws_demo_version.c file.
* Remove aws_demo_version.c and aws_application_version.h from the demo
  and test common CMakeLists.txt files.
* Remove the aws_application_version.h file and the xAppFirmwareVersion
  variable from port main.c files.
* Update preexisting ota_demo_config.h files to define the OTA firmware
  version macros that are used by the demos.
* Add the appFirmwareVersion declaration to the OTA common demo helpers
  file.
* Remove ota_appversion32.c.
* Remove aws_application_version.h.
* Update the Cypress OTA PAL port to get the firmware version number
  from the ota_config.h file instead of the aws_application_version.h
  file.

Update IDE projects to reflect the latest changes.

Only build the app version for the OTA demo
* The ota_demo_dependencies target was still being added to the project
when the tests are enabled. This change moves the dependency on the
ota_demo_version target to the demo cmake file to prevent this.

Remove references to the OTA library from the Cypress and Nuvoton
build files.