Restrict domain name length for Inventek module's DNS lookup function
The inventek WiFi module supports maximum domain name length of 64.
Trying to resolve a longer name leaves the module in a non-responsive
state. This PR adds a check to prevent calling ES_WIFI_DNS_LookUp for
longer names, thereby preventing the module from going into a
non-responsive state.

The following is from the datasheet of the module:
'D0' DNS Lookup:
This command performs a DNS lookup of a Domain Name to get its IPv4
address. The Domain Name is limited to 64 characters.

This PR also increases the stack size of test runner task to address a
stack overflow.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>