esp common_io uart: fix word length handling
The ioctl for configuring the UART was setting the data_bits member of
uart_config_t directly from the value in an IotUARTConfig_t ucWordlength
member. This is incorrect because the former is an enum and the latter
is an unsigned integer. Code for performing the translation has been
added.

Signed-off-by: Shubham Kulkarni <shubham.kulkarni@espressif.com>