Redirect network logging messages via logging task
Logging messages printed to the UART are serialized via logging task to
ensure that they are not mangled. Network logging messages were written
to UART directly which occasionally would result in mangled messages.

As a result CI would detect a test failure even though it succeded. This
commit redirects network logging message via logging task to ensure that
they are not mangled with other logging messages.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>