[demos/coreMQTT_Agent] set SOCKETS_SO_WAKEUP_CALLBACK
Use the SOCKETS_SO_WAKEUP_CALLBACK socket option to set a callback to be
executed when the callback is ready for reading.
prvMQTTClientSocketWakeupCallback() sends a PROCESSLOOP command to the
agent, which causes it to wake up and receive the new data immediately
instead of waiting until some other command arrives or
MQTT_AGENT_MAX_EVENT_QUEUE_WAIT_TIME elapses.